% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nneo-package.R
\docType{package}
\name{nneo-package}
\alias{nneo-package}
\alias{nneo}
\title{nneo - R client for the NEON API}
\description{
NEON API docs: \url{http://data.neonscience.org/data-api}
}
\section{nneo API}{

\itemize{
\item \code{\link[=nneo_products]{nneo_products()}}
\item \code{\link[=nneo_product]{nneo_product()}}
\item \code{\link[=nneo_sites]{nneo_sites()}}
\item \code{\link[=nneo_site]{nneo_site()}}
\item \code{\link[=nneo_locations]{nneo_locations()}}
\item \code{\link[=nneo_location]{nneo_location()}}
\item \code{\link[=nneo_data]{nneo_data()}}
\item \code{\link[=nneo_file]{nneo_file()}}
\item \code{\link[=nneo_wrangle]{nneo_wrangle()}}
}
}

\section{Curl options}{

Curl options follow those in the \pkg{curl} package. Find the curl
options by looking at \code{\link[curl:curl_options]{curl::curl_options()}}. Pass them
on into any \pkg{nneo} function call.
}

\section{User-Agent}{

A user agent string is sent in every request just to let the NEON
servers know that the request is coming from R and from this package.
No personal info is shared, just the package name and version for
\pkg{nneo} and the underlying HTTP packages (\pkg{crul}, \pkg{curl})
}

\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}

Josh Roberti \email{jaroberti87@gmail.com}
}
\keyword{package}
