% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nneo_locations.R
\name{nneo_locations}
\alias{nneo_locations}
\alias{nneo_location}
\title{Locations}
\usage{
nneo_locations(...)

nneo_location(x, ...)
}
\arguments{
\item{...}{named curl options passed on to \code{\link[crul]{HttpClient}},
e.g., \code{list(verbose = TRUE)} or \code{list(timeout_ms = 1000)}}

\item{x}{(character) a location code name}
}
\value{
a tibble (data.frame)
}
\description{
Locations
}
\examples{
\dontrun{
## list products
nneo_locations()

## get a product
res <- nneo_location("HARV")
res
res$locationProperties
res$locationChildren
}
}
