% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nneo_products.R
\name{nneo_products}
\alias{nneo_products}
\alias{nneo_product}
\title{Products}
\usage{
nneo_products(...)

nneo_product(x, ...)
}
\arguments{
\item{...}{named curl options passed on to \code{\link[crul]{HttpClient}},
e.g., \code{list(verbose = TRUE)} or \code{list(timeout_ms = 1000)}}

\item{x}{(character) a product code}
}
\value{
\code{nneo_products} returns a tibble (data.frame), and
\code{nneo_product} returns a list
}
\description{
Products
}
\examples{
\dontrun{
## list products
nneo_products()

## get a product
res <- nneo_product("DP3.30018.001")
res$productDescription

### get many products
ids <- c("DP3.30018.001", "DP4.00001.001", "DP3.30025.001")
lapply(ids, nneo_product)

## curl options
nneo_product("DP3.30018.001", verbose = TRUE)
nneo_product("DP3.30018.001", verbose = TRUE, timeout_ms = 1000)
nneo_product("DP3.30018.001", verbose = TRUE, useragent = "hello world")
}
}
