% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nneo_sites.R
\name{nneo_sites}
\alias{nneo_sites}
\alias{nneo_site}
\title{Sites}
\usage{
nneo_sites(...)

nneo_site(x, ...)
}
\arguments{
\item{...}{named curl options passed on to \code{\link[crul]{HttpClient}},
e.g., \code{list(verbose = TRUE)} or \code{list(timeout_ms = 1000)}}

\item{x}{(character) a site code}
}
\value{
a tibble (data.frame)
}
\description{
Sites
}
\examples{
\dontrun{
## list products
nneo_sites()

## get a product
res <- nneo_site("JORN")
res$dataProducts
res$siteDescription
}
}
