% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nneo_sites.R
\name{nneo_sites}
\alias{nneo_site}
\alias{nneo_sites}
\title{Sites}
\usage{
nneo_sites(...)

nneo_site(x, ...)
}
\arguments{
\item{...}{curl options passed on to \code{\link[httr]{GET}}}

\item{x}{(character) a site code}
}
\value{
a tibble (data.frame)
}
\description{
Sites
}
\examples{
\dontrun{
## list products
nneo_sites()

## get a product
res <- nneo_site("JORN")
res$dataProducts
res$siteDescription
}
}

